Cleanup - rename classes
- Tile -> SimpleTile
- ITile -> Tile
